+++ Sources/TorliStats/StatusBar/TorliAppDelegate+StatusBar.swift
@@ -33,9 +33,11 @@ extension TorliAppDelegate {
             guard let self else { return }
             self.updateStatusBarRunnerImage(image)
         }
+        statusLogoAnimator?.setPaused(monitoringPauseController.mode != .realtime)
     }
 
     func updateStatusBarLogoSpeed() {
+        statusLogoAnimator?.setPaused(monitoringPauseController.mode != .realtime)
         statusLogoAnimator?.setCPUUsage(store.cpu)
     }